fix sscanf bug failing to detect value item in instrument spec
%255c means "read exactly 255 chars", not "read up to 255 chars".

according to C spec 7.21.6.2 "The fscanf function" ¶12

> Matches a sequence of characters of exactly the number specified by the
> field width (1 if no field width is present in the directive).

this is a longstanding bug in glibc:

https://sourceware.org/bugzilla/show_bug.cgi?id=12701

which probably was the cause for the author of the code not noticing it.
it fails as it should with musl libc, which results in parsing errors like:
"there should be 12 instruments, but only 11 are found!".
the misleading message has also been tweaked, as there are actually only
11 instruments plus a value.

we therefore use %s and assume the 256 byte buffer size is sufficient to
parse the remainder of the line, as it is user-provided this should not
be a security issue.